Create README.md
It is a game made with Scratch.
In short, it is about a shark that you have to move up or down using the keyboard so that it catches the fish that moves in its direction. The game ends when one of these two situations occurs:
- catch 5 fish
- fails 3 times
The messages at the end are different.
In the game there are several sprites that will be displayed in the scene and have movements and costume changes. Sounds and changes of scenery are handled.